Add example which demonstrates finding an eigenfunction and eigenvalue.
Implement Schrodingers equation for a 1D particle-in-a-box. Include a loss term for the inner product with another function (sine in this case), as a method of finding new eigenfunctions once existing ones have been found. Default configuration includes modes 1 & 2 as orthogonality functions, so should learn 3rd mode (sin 3 pi x).

Implement IntegralInteriorConstraint; very similar to IntegralBoundaryConstraint but for the interior region of geometries. Used for normalization constraint of wavefunction.

Signed-off-by: Jacob Silterra <github@jacobsilterra.com>